+++ Sources/Models/ALKConfiguration.swift
@@ -195,6 +195,7 @@ public struct ALKConfiguration {
 
     public var conversationInfoModel : KMConversationInfoViewModel?
     
+    /// if true then delete option will be enabled for end user on long press, By default it is false
     public var enableDeleteConversationOnLongpress : Bool = false
 
     /// If true, contact share option in chatbar will be hidden.