+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
-## [Unreleased]
+## [1.0.6] 2023-04-19
 - Added Support for delete conversation for End User
 - Fixed Form text area data not added with message for `postFormDataAsMessage` issue
 - Matched form action flow with web & android.